Keep two versions for backwards compatability
